requirements.txt
# ═══════════════════════════════════════════════════════════════
# SIMPLE MESSAGE XCHANGER - Dependências
# Autor: Elias Andrade - Replika AI - Maringá/PR
# ═══════════════════════════════════════════════════════════════

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
cryptography>=42.0.0
PyJWT>=2.8.0
requests>=2.31.0
PyYAML>=6.0.0
